fixed conversion warning
Former-commit-id: 2345e71bebfb2c7c6ec156e3c1ed91cf88639dd2

+++ src/automat.cpp
@@ -156,7 +156,7 @@ unsigned int Automat::getNeighborsOfType(const size_t x, const size_t y, const s
 			if (overflowEdges) new_y = 0;
 			else continue;
 		}
-		if (getCellTypeAt(new_x, new_y) == cellType) {
+		if (getCellTypeAt((size_t)new_x, (size_t)new_y) == cellType) {
 			count++;
 		}
 	}
